ValidateServiceLogLevelOrPanic moved to run.go
