+++ cmd/devtools/run.go
@@ -330,3 +330,16 @@ func getCometbftOKCallback(config []string) func() bool {
 		}
 	}
 }
+
+// ValidateServiceLogLevelOrPanic validates the service log level and panics if
+// it is invalid. The valid log levels are: debug, info, error.
+func ValidateServiceLogLevelOrPanic(logLevel string) {
+	switch logLevel {
+	case "debug", "info", "error":
+		return
+	default:
+		log.WithField("service-log-level", logLevel).Fatal("Invalid service log level. Must be one of: 'debug', 'info', 'error'")
+		panic("Invalid service log level")
+	}
+
+}
